backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
psycopg2==2.9.5
django-cors-headers==3.14.0
celery>=5.2,<6.0
redis>=4.2
